'Tis 1 in the morning and i have finally finished the conversion between eth/btc for the balance change.
It now shows the change in terms of BTC which calculates the price difference compared to if all of your transactions were buying BTC/ETH instead of the transactions currency.
This is a much more effective way of informing the user whether they are 'up'/'down' in BTC/ETH over a time period upposed to the lazy way of just converting the usd change to btc/eth. Anyone can do that... This is actually useful!

CBA to write what i did... Summary i got the prices of all the currencies at the time frame along with btc/eth. Made some comparisons with the current prices and then calculated the difference between the two.

Also, price grabe for fiat/crypto create/update were wrong, fixed them :)